feat: Member entity
